# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sqordia API",
    "description": "Comprehensive API for business plan generation and management platform with Google OAuth authentication support",
    "version": "1.0.0",
    "contact": {
      "name": "Sqordia API Support",
      "email": "support@sqordia.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.sqordia.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/obnl/plans": {
      "post": {
        "tags": ["OBNL Business Plans"],
        "summary": "Create OBNL business plan",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOBNLPlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OBNL plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OBNLPlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/obnl/plans/{planId}": {
      "get": {
        "tags": ["OBNL Business Plans"],
        "summary": "Get OBNL plan details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OBNL plan details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OBNLPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates": {
      "get": {
        "tags": ["Templates"],
        "summary": "Get all templates",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Templates retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Templates"],
        "summary": "Create new template",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OBNLPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "mission": {
            "type": "string"
          },
          "vision": {
            "type": "string"
          },
          "obnlType": {
            "type": "string"
          },
          "targetAudience": {
            "type": "string"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fundingSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "complianceRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "complianceStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateOBNLPlanRequest": {
        "type": "object",
        "required": ["organizationId", "mission", "vision", "obnlType"],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "mission": {
            "type": "string"
          },
          "vision": {
            "type": "string"
          },
          "obnlType": {
            "type": "string"
          },
          "targetAudience": {
            "type": "string"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fundingSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "complianceRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "OBNLPlanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "TemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "audience": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "usageCount": {
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTemplateRequest": {
        "type": "object",
        "required": ["name", "description", "category"],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "audience": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}